feat(log_utils): Write in GitHub Summary
[release:minor]

Closes #8
