+++ .github/workflows/python-publish.yml
@@ -11,6 +11,10 @@ name: Upload Python Package
 on:
   release:
     types: [published]
+  workflow_run:
+    workflows: ["Release"]
+    branches: [master]
+    types: [completed]
 
 permissions:
   contents: read
